requirements.txt
opencv-python
face-recognition
dlib==19.24.4
Pillow
pyttsx3